Model loading improvements
Fixed minor bug in texture loading where if the texture already existed then a null pointer would be returned instead of the texture.
Changed model loader to match texture loader functionality.
Added an exit flag for easier debugging, causes the app to exit gracefully at the end of the frame when set to true.